requirements.txt
apache-airflow
selenium
pandas
snowflake-connector-python
snowflake-snowpark-python
psycopg2-binary
python-dotenv
webdriver-manager